third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/elements-panel-selection-after-delete-expected.txt

Tests that elements panel correctly updates selection on node removal.


Running: testUserDelete
Selecting node...
The following node is now selected: <DIV class="child2">
Deleting selected node like it was a user action...
The following node is now selected: <DIV class="child3">
Deleting selected node like it was a user action again...
The following node is now selected: <DIV class="child1">
Deleting last child like it was a user action...
The following node is now selected: <DIV class="left">

Running: testExternalDelete
Selecting node...
The following node is now selected: <DIV class="child2">
Deleting selected node externally...
The following node is now selected: <DIV class="child3">
Deleting selected node externally again...
The following node is now selected: <DIV class="child1">
Deleting last child externally...
The following node is now selected: <DIV class="left">
